+++ NEWS.md
@@ -1,5 +1,9 @@
 # cli (development version)
 
+* cli new replaces newline characters within `{.class ... }` inline styles
+  with spaces. If the `cli.warn_inline_newlines` option is set to FALSE, then
+  it also throws a warning. (#417).
+
 * `code_highlight` now falls to the default theme (instead of no theme) for
   unknown RStudio themes (#482, @rossellhayes).